WIP: Refactor towards a free monad architecture.
The logic is contained in two interpreters. The first interprets the
client language, with terms such as 'visit' and 'buy' in terms of the
game language, with terms such as 'transfer' and 'computeRent'. The
second interprets the game language into some monad that keeps the game
state; here we've used SqlPersistT.